eyesmedia-version-info-services/requirements.txt
boto3==1.9.33
botocore==1.12.33
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==3.0
docutils==0.16
elasticsearch==6.3.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.0
numpy==1.19.1
pandas==1.0.5
peewee==3.13.3
pycparser==2.20
PyJWT==1.7.1
pymongo==3.7.2
PyMySQL==0.9.2
python-dateutil==2.7.5
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.1.13
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
eyesmediapydb>=0.9.2
eyesmediapyutils>=0.9.2